+++ layers/Engine/packages/migrate-component-model/migrate/initialization.php
@@ -7,12 +7,6 @@ class Initialization
     {
         load_plugin_textdomain('pop-engine', false, dirname(plugin_basename(__FILE__)).'/languages');
 
-        /**
-         * Load the Server first, so we can access class ServerUtils
-         * And its required library first
-         */
-        include_once 'server/load.php';
-
         // /**
         //  * Load the Contracts
         //  */
