+++ layers/Engine/packages/migrate-component-model/migrate/kernel/library/js-keys.php
@@ -1,6 +1,6 @@
 <?php
 
 // Make sure each JS Key is unique
-define('GD_JS_MODULE', \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys() ? 'm' : 'module');
-define('GD_JS_SUBMODULES', \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys() ? 'ms' : 'submodules');
-define('GD_JS_MODULEOUTPUTNAME', \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys() ? 's' : 'moduleoutputname');
+define('GD_JS_MODULE', \PoP\ComponentModel\Environment::compactResponseJsonKeys() ? 'm' : 'module');
+define('GD_JS_SUBMODULES', \PoP\ComponentModel\Environment::compactResponseJsonKeys() ? 'ms' : 'submodules');
+define('GD_JS_MODULEOUTPUTNAME', \PoP\ComponentModel\Environment::compactResponseJsonKeys() ? 's' : 'moduleoutputname');
