--- layers/Engine/packages/migrate-component-model/migrate/server/configuration/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'server-utils.php';
