--- layers/Engine/packages/migrate-component-model/migrate/server/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'configuration/load.php';
