+++ layers/Schema/packages/migrate-everythingelse/migrate/plugins/pop-coreprocessors/kernel/library/js-keys.php
@@ -1,7 +1,7 @@
 <?php
 use PoP\Hooks\Facades\HooksAPIFacade;
 
-$compact = \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys();
+$compact = \PoP\ComponentModel\Environment::compactResponseJsonKeys();
 define('GD_JS_FONTAWESOME', $compact ? 'fa' : 'fontawesome');
 
 HooksAPIFacade::getInstance()->addFilter('gd_jquery_constants', 'wassupJqueryConstantsJsparams');
