+++ layers/Schema/packages/migrate-everythingelse/migrate/plugins/pop-engine-htmlcssplatform/kernel/library/js-keys.php
@@ -1,7 +1,7 @@
 <?php
 use PoP\Hooks\Facades\HooksAPIFacade;
 
-$compact = \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys();
+$compact = \PoP\ComponentModel\Environment::compactResponseJsonKeys();
 define('GD_JS_TEMPLATE', $compact ? 't' : 'template');
 define('GD_JS_SUBMODULEOUTPUTNAMES', $compact ? 'ss' : 'submoduleoutputnames');
 define('POP_JS_TEMPLATES', $compact ? 'tx' : 'templates');
