+++ layers/Schema/packages/migrate-everythingelse/migrate/plugins/pop-engine-webplatform/library/functions/constants.php
@@ -3,8 +3,8 @@ use PoP\Hooks\Facades\HooksAPIFacade;
 
 define('POP_PARAMS_PARAMSSCOPE_URL', 'paramsscope-url');
 
-define('POP_PROGRESSIVEBOOTING_CRITICAL', \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys() ? 'c' : 'critical');
-define('POP_PROGRESSIVEBOOTING_NONCRITICAL', \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys() ? 'n' : 'noncritical');
+define('POP_PROGRESSIVEBOOTING_CRITICAL', \PoP\ComponentModel\Environment::compactResponseJsonKeys() ? 'c' : 'critical');
+define('POP_PROGRESSIVEBOOTING_NONCRITICAL', \PoP\ComponentModel\Environment::compactResponseJsonKeys() ? 'n' : 'noncritical');
 
 HooksAPIFacade::getInstance()->addFilter('gd_jquery_constants', 'popWebPlatformJqueryConstantsImpl');
 function popWebPlatformJqueryConstantsImpl($jqueryConstants)
