+++ layers/Schema/packages/migrate-everythingelse/migrate/plugins/pop-resourceloader/library/functions/constants.php
@@ -1,5 +1,5 @@
 <?php
 
-$compact = \PoP\ComponentModel\Server\Utils::compactResponseJsonKeys();
+$compact = \PoP\ComponentModel\Environment::compactResponseJsonKeys();
 define('GD_JS_RESOURCES', $compact ? 'r' : 'resources');
 define('GD_SEPARATOR_RESOURCELOADER', '|');
