+++ layers/Schema/packages/migrate-everythingelse/migrate/plugins/pop-serviceworkers/plugins/pop-resourceloader/library/functions/hooks.php
@@ -16,7 +16,7 @@ class PoP_ServiceWorkers_ResourceLoaderProcessor_Hooks
 
         // Comment Leo 14/11/2017: enqueue scripts only if we enable SW by configuration,
         // or if allowing for the config so it does not explode due to the different resources in the bundlegroup
-        if (\PoP\ComponentModel\Server\Utils::enableConfigByParams() || !PoP_ServiceWorkers_ServerUtils::disableServiceworkers()) {
+        if (\PoP\ComponentModel\Environment::enableConfigByParams() || !PoP_ServiceWorkers_ServerUtils::disableServiceworkers()) {
             // Comment Leo 14/11/2017: we must execute /sw.js before executing /sw-registrar.js
             // Otherwise, when first loading the page, if it has changed, it will not show the message "Please click here to reload the page",
             // since the code to execute that is executed after the SW 'refresh' function...
